callcenter cost

+++ app-call-center/src/main/java/com/lsxy/call/center/dao/CallCenterDao.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 呼叫中心系统
@@ -16,5 +17,5 @@ public interface CallCenterDao extends BaseDaoInterface<CallCenter, Serializable
 
     @Modifying
     @Query("update CallCenter a set a.cost=a.cost+:cost where a.id=:id")
-    public void incrCost(@Param("id") String id,@Param("cost") double cost);
+    public void incrCost(@Param("id") String id,@Param("cost") BigDecimal cost);
 }
